Add an '=' to the example Using 'if...else'
This will add an extra '=' to the example in Example Using "if...else". While, the comments indicate that x = 5, making x an integer, it is better to use triple '=' for equality without type coercion so mistakes do not occur if 'x' were happen to be set to the string 'x'.